refactor(ai): extract tool definitions to tool_defs.rs to satisfy 1000-line limit
